Move hardcoded data, rename some screen composables
