+++ ui-components/src/commonMain/kotlin/org/jetbrains/kotlinconf/ui/components/PartnerCard.kt
